requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.1.1
fastapi==0.110.0
filelock==3.15.4
flatbuffers==23.5.26
fsspec==2024.6.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.6
imageio==2.34.2
itsdangerous==2.1.2
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
orjson==3.9.15
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyclipper==1.3.0.post5
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
# pytesseract==0.3.10
python-bidi==0.4.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.36.3
sympy==1.12
tifffile==2024.7.2
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0